% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Translation Vector Function}
\usage{
center(mean_data, sensor_idx)
}
\arguments{
\item{mean_data}{A 2D matrix representing the mean points in the X, Y, and Z dimensions of three sensors}

\item{sensor_idx}{A vector of the numeric ids of the three plane sensors}
}
\value{
A vector with a length of three representing the translation vector for rotation
}
\description{
Calculates the translation vector for rotation
}
