% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_mov.R
\name{correct_mov}
\alias{correct_mov}
\title{Correct Movement Function (correct_mov)}
\usage{
correct_mov(filtered, coord, ref_idx, base_rt, base_center)
}
\arguments{
\item{filtered}{A 3d array of interpolated and filtered data from the data recording}

\item{coord}{A 3d array of the rotated bite plane data}

\item{ref_idx}{A vector of the numeric ids of the three referent sensors}

\item{rotation}{A rotation matrix extracted from define_coord}

\item{center}{A vector with a length of 3 representing the translation vector extracted from define_coord}
}
\value{
A 3D array of the rotated data from the data recording
}
\description{
Purpose: This function rotates 3D articulatory kinematic data from a data recording
in reference to rotated bite plane data.
}
