% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_coord.R
\name{define_coord}
\alias{define_coord}
\title{Define Coordinate Plane Function}
\usage{
define_coord(data, ref_idx, bp_idx, flip_axis = FALSE)
}
\arguments{
\item{data}{A string representing the name of the data matrix (imported from load_tsv)}

\item{ref_idx}{A vector of the numeric ids of the three referent sensors}

\item{bp_idx}{A vector of the numeric ids of the bite plane sensors}

\item{flip_axis}{A optional conditional argument that flips the referent vector for purposes of orientation}
}
\value{
A list containing the 3D array of the rotated coordinate plane, the rotation matrix
and translation vector
}
\description{
Purpose: This function rotates 3D articulatory kinematic data from a bite plan recording
to redefine a coordinate plane for movement data.
}
