% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_filter.R
\name{interp_filter}
\alias{interp_filter}
\title{Interpolation and Filtering of EMA Data}
\usage{
interp_filter(raw, ref_idx)
}
\arguments{
\item{ref_idx}{A vector representing the numeric ids of the three referent sensors}

\item{data}{A string representing the name of the data matrix containing the raw EMA recording.}
}
\value{
The resulting 3D matrix of data
}
\description{
Purpose: This function finds missing values in an EMA recording, uses linear 
interpolation to fill in the missing values. And then smooths the data using a 
low-pass 20 Hz butterworth filter.
}
