% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tsv.R
\name{load_tsv}
\alias{load_tsv}
\title{Load TSV Data}
\usage{
load_tsv(file)
}
\arguments{
\item{file}{A string representing the path to the TSV file}
}
\value{
A list containing a 3D matrix of data and a numeric vector of timestamps
}
\description{
Purpose: This function loads a TSV file into R, processes the data, and returns a list containing the 
3D matrix and a vector of timestamps.
}
