% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_matrix.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation Matrix Function}
\usage{
rotation_matrix(axis, angle)
}
\arguments{
\item{axis}{A numeric vector of 3 representing the axis of rotation}

\item{angle}{A numeric value representing the angle of rotation (radians)}
}
\value{
A 3x3 rotation matrix as a numeric matrix
}
\description{
Constructs a 3D rotation matrix given an axis of rotation and an angle.
}
