% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_palate.R
\name{est_palate}
\alias{est_palate}
\title{Palate trace Function (est_palate)}
\usage{
est_palate(data, coord, ref_idx, pl_idx, base_rt, base_center)
}
\arguments{
\item{data}{A 3d array of the palate trace recording}

\item{coord}{A 3D array of the rotated coordinate plane}

\item{pl_idx}{The numeric value representing the palate trace sensor}

\item{base_rt}{The rotation matrix extracted from the rotated coordinate plane}

\item{base_center}{A vector with a length of 3 representing the translation vector extracted from the rotated coordinate plane}
}
\value{
A data frame of the estimated X, Y, and Z coordinates of the palate
}
\description{
Purpose: This function rotates 3D articulatory kinematic data from a palate tracing,
rotates and corrects the data from head movement, and then calculates an estimated palate location 
in the XYZ plane.
}
