% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vec.R
\name{norm_vec}
\alias{norm_vec}
\title{Normal Vector Calculation Function}
\usage{
norm_vec(mean_data, sensor_idx)
}
\arguments{
\item{mean_data}{A 2D matrix representing the mean points in the X, Y, and Z dimensions of three sensors}

\item{sensor_idx}{A vector of the numeric ids of the three plane sensors}
}
\value{
A vector with a length of three representing the normal vector of the plane
}
\description{
Calculates the normal vector in reference to a plane defined by three points
}
