% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate Sensor Recording Function (rotate)}
\usage{
rotate(data, coord, ref_idx, base_rt, base_center, time)
}
\arguments{
\item{data}{A 3d array of the sensor recording data}

\item{coord}{A 3d array of the rotated bite plane data}

\item{ref_idx}{A vector of the numeric ids of the three referent sensors}

\item{time}{A vector representing the time stamps of the original sensor data recording}

\item{rotation}{A rotation matrix extracted from define_coord}

\item{center}{A vector with a length of 3 representing the translation vector extracted from define_coord}
}
\value{
A data frame of the rotated data ready for analysis
}
\description{
This function rotates 3D articulatory kinematic data from a  recording and 
converts it into a dataframe ready for analysis.
}
